image/test/mochitest/test_animated_gif.html

<!DOCTYPE HTML>
<html>
<head>
<meta charset="utf-8">
<title>Images outside of display port are not decoded</title>
<script src="/tests/SimpleTest/SimpleTest.js"></script>
<script src="imgutils.js"></script>
<link rel="stylesheet" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<p id="display"></p>
<div id="content" style="display: none"></div>
<div id="scroller" style="height: 300px; overflow: scroll;">
  <div style="width: 100%; height: 5000px;"></div>
  <iframe src="http://example.org/tests/image/test/mochitest/child.html"></iframe>
  <div style="width: 100%; height: 5000px;"></div>
</div>
<pre id="test"></pre>
<script>
  SimpleTest.waitForExplicitFinish();

  let loadCompleteCount = 0;
  window.addEventListener("message", event => {
    isnot(event.data, "decodeComplete",
          "The image outside of the display port should not be decoded");
    if (event.data == "loadComplete") {
      loadCompleteCount++
    }
  });

  const start = document.timeline.currentTime;
  // Waits a second;
  SimpleTest.waitForCondition(() => {
    return 1000 < (document.timeline.currentTime - start);
  }, () => {
    is(loadCompleteCount, 1,
       "`loadComplete` should be received once");
    SimpleTest.finish();
  });
</script>
</body>
</html>
